glens/worker/object-detecgion-worker/requirements.txt
absl-py==0.10.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
future==0.18.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
importlib-metadata==2.0.0
kiwisolver==1.2.0
Markdown==3.3.1
matplotlib==3.3.2
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.2
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.3.0
